requirements.txt
tensorflow==2.12.0
pandas 
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .
